Analysis: Nest the render graph builder under RenderGraph
